Added option to display dynamic variables menu when no menu is present
