+++ src/components/Editor/Menu/DynamicVariables.jsx
@@ -1,0 +1,57 @@
+import DynamicVariables from "neetomolecules/DynamicVariables";
+import { useTranslation } from "react-i18next";
+
+const MenuDynamicVariables = ({
+  editor,
+  variables = [],
+  standalone = false,
+}) => {
+  const { t } = useTranslation();
+  const handleVariableClick = item => {
+    const { category, categoryLabel, key, label } = item;
+    const variableName = category ? `${category}.${key}` : key;
+    const variableLabel = category
+      ? `${categoryLabel || category}:${label}`
+      : label;
+
+    editor
+      .chain()
+      .focus()
+      .setVariable({ id: variableName, label: variableLabel })
+      .run();
+  };
+
+  return (
+    <div
+      className="neeto-editor-fixed-menu__variables"
+      data-cy="neeto-editor-fixed-menu-variables"
+      style={
+        standalone
+          ? {
+              position: "absolute",
+              right: "0px",
+              top: "8px",
+              bottom: "auto",
+              zIndex: 10,
+            }
+          : undefined
+      }
+    >
+      <DynamicVariables
+        {...{ variables }}
+        dropdownProps={{
+          buttonSize: "small",
+          buttonProps: {
+            tooltipProps: {
+              content: t("neetoEditor.menu.dynamicVariables"),
+              position: "bottom",
+            },
+          },
+        }}
+        onVariableClick={handleVariableClick}
+      />
+    </div>
+  );
+};
+
+export default MenuDynamicVariables;
