+++ src/components/Editor/Menu/Fixed/index.jsx
@@ -3,9 +3,6 @@ import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import classNames from "classnames";
 import { EDITOR_OPTIONS } from "common/constants";
 import { isNotEmpty } from "neetocist";
-import DynamicVariables from "neetomolecules/DynamicVariables";
-import { isEmpty } from "ramda";
-import { useTranslation } from "react-i18next";
 
 import MediaUploader from "components/Editor/MediaUploader";
 
@@ -14,6 +11,8 @@ import MoreMenu from "./components/MoreMenu";
 import { MENU_ELEMENTS } from "./constants";
 import useEditorState from "./hooks/useEditorState";
 import { reGroupMenuItems, buildMenuOptions } from "./utils";
+
+import MenuDynamicVariables from "../DynamicVariables";
 
 const Fixed = ({
   editor,
@@ -48,8 +47,6 @@ const Fixed = ({
 
   const menuRef = useRef(null);
   const menuContainerRef = useRef(null);
-
-  const { t } = useTranslation();
 
   const runEditorCommand = useCallback(
     command => () => command(editorRef.current),
@@ -137,20 +134,6 @@ const Fixed = ({
     options.includes(EDITOR_OPTIONS.IMAGE_UPLOAD) ||
     options.includes(EDITOR_OPTIONS.VIDEO_UPLOAD);
 
-  const handleVariableClick = item => {
-    const { category, categoryLabel, key, label } = item;
-    const variableName = category ? `${category}.${key}` : key;
-    const variableLabel = category
-      ? `${categoryLabel || category}:${label}`
-      : label;
-
-    editor
-      .chain()
-      .focus()
-      .setVariable({ id: variableName, label: variableLabel })
-      .run();
-  };
-
   return (
     <div
       ref={menuContainerRef}
@@ -182,26 +165,7 @@ const Fixed = ({
           <MoreMenu {...{ editor }} groups={moreMenuItems} />
         )}
       </div>
-      {!isEmpty(variables) && (
-        <div
-          className="neeto-editor-fixed-menu__variables"
-          data-cy="neeto-editor-fixed-menu-variables"
-        >
-          <DynamicVariables
-            {...{ variables }}
-            dropdownProps={{
-              buttonSize: "small",
-              buttonProps: {
-                tooltipProps: {
-                  content: t("neetoEditor.menu.dynamicVariables"),
-                  position: "bottom",
-                },
-              },
-            }}
-            onVariableClick={handleVariableClick}
-          />
-        </div>
-      )}
+      <MenuDynamicVariables {...{ editor, variables }} />
       {isAddLinkActive && (
         <LinkAddPopOver
           {...{
